+++ pianoKiet.h
@@ -6,6 +6,6 @@
 using namespace std;
 
 //Insertion Sort
-void insertionSort_countCompare(int* a, int n, int& count_compare);
-void insertionSort_countTime(int* a, int n, int& count_compare);
-void shakerSort_countTime(int* a, int n, int& count_compare);
+void insertionSort_countCompare(int* a, int n, long long& count_compare);
+void insertionSort_countTime(int* a, int n, long long& count_compare);
+void shakerSort_countTime(int* a, int n, long long& count_compare);
